+++ relay-server/src/services/store.rs
@@ -963,8 +963,8 @@ impl StoreService {
 
         let d = &mut Deserializer::from_slice(&payload);
 
-        let mut log: LogKafkaMessage = match serde_path_to_error::deserialize(d) {
-            Ok(log) => log,
+        let logs: LogKafkaMessages = match serde_path_to_error::deserialize(d) {
+            Ok(logs) => logs,
             Err(error) => {
                 relay_log::error!(
                     error = &error as &dyn std::error::Error,
@@ -992,37 +992,42 @@ impl StoreService {
             }
         };
 
-        log.organization_id = scoping.organization_id.value();
-        log.project_id = scoping.project_id.value();
-        log.retention_days = retention_days;
-        log.received = safe_timestamp(received_at);
+        for mut log in logs.items {
+            log.organization_id = scoping.organization_id.value();
+            log.project_id = scoping.project_id.value();
+            log.retention_days = retention_days;
+            log.received = safe_timestamp(received_at);
 
-        let message = KafkaMessage::Log {
-            headers: BTreeMap::from([("project_id".to_string(), scoping.project_id.to_string())]),
-            message: log,
-        };
+            let message = KafkaMessage::Log {
+                headers: BTreeMap::from([(
+                    "project_id".to_string(),
+                    scoping.project_id.to_string(),
+                )]),
+                message: log,
+            };
 
-        self.produce(KafkaTopic::OurLogs, message)?;
+            self.produce(KafkaTopic::OurLogs, message)?;
 
-        // We need to track the count and bytes separately for possible rate limits and quotas on both counts and bytes.
-        self.outcome_aggregator.send(TrackOutcome {
-            category: DataCategory::LogItem,
-            event_id: None,
-            outcome: Outcome::Accepted,
-            quantity: 1,
-            remote_addr: None,
-            scoping,
-            timestamp: received_at,
-        });
-        self.outcome_aggregator.send(TrackOutcome {
-            category: DataCategory::LogByte,
-            event_id: None,
-            outcome: Outcome::Accepted,
-            quantity: payload_len as u32,
-            remote_addr: None,
-            scoping,
-            timestamp: received_at,
-        });
+            // We need to track the count and bytes separately for possible rate limits and quotas on both counts and bytes.
+            self.outcome_aggregator.send(TrackOutcome {
+                category: DataCategory::LogItem,
+                event_id: None,
+                outcome: Outcome::Accepted,
+                quantity: 1,
+                remote_addr: None,
+                scoping,
+                timestamp: received_at,
+            });
+            self.outcome_aggregator.send(TrackOutcome {
+                category: DataCategory::LogByte,
+                event_id: None,
+                outcome: Outcome::Accepted,
+                quantity: payload_len as u32,
+                remote_addr: None,
+                scoping,
+                timestamp: received_at,
+            });
+        }
 
         Ok(())
     }
@@ -1388,6 +1393,16 @@ struct CheckInKafkaMessage {
 }
 
 #[derive(Debug, Deserialize, Serialize)]
+struct SpanLink<'a> {
+    pub trace_id: &'a str,
+    pub span_id: &'a str,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub sampled: Option<bool>,
+    #[serde(borrow)]
+    pub attributes: Option<&'a RawValue>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
 struct SpanMeasurement {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     value: Option<f64>,
@@ -1413,6 +1428,8 @@ struct SpanKafkaMessage<'a> {
     data: Option<&'a RawValue>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     kind: Option<&'a str>,
+    #[serde(default, skip_serializing_if = "none_or_empty_vec")]
+    links: Option<Vec<SpanLink<'a>>>,
     #[serde(borrow, default, skip_serializing_if = "Option::is_none")]
     measurements: Option<BTreeMap<Cow<'a, str>, Option<SpanMeasurement>>>,
     #[serde(default)]
@@ -1489,6 +1506,12 @@ struct LogAttribute {
 }
 
 #[derive(Debug, Deserialize, Serialize)]
+struct LogKafkaMessages<'a> {
+    #[serde(borrow)]
+    items: Vec<LogKafkaMessage<'a>>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
 struct LogKafkaMessage<'a> {
     #[serde(default)]
     organization_id: u64,
@@ -1525,6 +1548,13 @@ fn none_or_empty_object(value: &Option<&RawValue>) -> bool {
     match value {
         None => true,
         Some(raw) => raw.get() == "{}",
+    }
+}
+
+fn none_or_empty_vec<T>(value: &Option<Vec<T>>) -> bool {
+    match &value {
+        Some(vec) => vec.is_empty(),
+        None => true,
     }
 }