+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+**Features**:
+
+- Custom attachment expansion for Switch. ([#4566](https://github.com/getsentry/relay/pull/4566))
+
+## 25.4.0
+
 - Extract searchable context fields into sentry tags for segment spans. ([#4651](https://github.com/getsentry/relay/pull/4651))
 
 **Features**:
@@ -43,6 +49,7 @@
 - Add `/v1/traces` (without a trailing slash) as a spec-compliant alternative for our OTLP traces endpoint. ([#4655](https://github.com/getsentry/relay/pull/4655))
 - Improve handling of failed Redis connections. ([#4657](https://github.com/getsentry/relay/pull/4657))
 - Expose new metrics from the async pool. ([#4658](https://github.com/getsentry/relay/pull/4658))
+- Serialize span's `links` information when producing to Kafka. ([#4601](https://github.com/getsentry/relay/pull/4601))
 
 ## 25.3.0